NIFI-305: Refactoring superclass BinFiles from MergeContent
